+++ internal/applets/applet.go
@@ -22,6 +22,7 @@ import (
 	"mimixbox/internal/applets/fileutils/mkdir"
 	"mimixbox/internal/applets/fileutils/mv"
 	"mimixbox/internal/applets/fileutils/rm"
+	"mimixbox/internal/applets/fileutils/rmdir"
 	"mimixbox/internal/applets/fileutils/touch"
 	"mimixbox/internal/applets/jokeutils/fakemovie"
 	"mimixbox/internal/applets/shellutils/chroot"
@@ -63,6 +64,7 @@ func init() {
 		"mv":        {mv.Run, "Rename SOURCE to DESTINATION, or move SOURCE(s) to DIRECTORY"},
 		"path":      {path.Run, "Manipulate filename path"},
 		"rm":        {rm.Run, "Remove file(s) or directory(s)"},
+		"rmdir":     {rmdir.Run, "Remove directory"},
 		"serial":    {serial.Run, "Rename the file to the name with a serial number"},
 		"sh":        {mbsh.Run, "Mimix Box Shell"},
 		"touch":     {touch.Run, "Update the access and modification times of each FILE to the current time"},