+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 # Changelog
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-## [0.4.0] - 2021-11-18
+## [0.5.1] - 2021-11-18
 ### Added
  - rm command.
+ - rmdir command.
+### Changed
+ - The library that was open to the public(pkg) to the internal library.
+### Fixed
+ - Bug that misjudged applet arguments as mimixbox arguments.
 ## [0.3.0] - 2021-11-17
 ### Added
  - touch command.